requirements.txt
streamlit
python-dotenv
beautifulsoup4
requests
youtube-transcript-api